migrate schema types
