+++ packages/core/src/schema/zod/android/configPlatformAndroid.ts
