+++ packages/core/src/schema/zod/android/configPlatformSharedAndroid.ts
