+++ packages/core/src/schema/zod/android/configPluginAndroid.ts
