+++ packages/core/src/schema/zod/configCommon.ts
