+++ packages/core/src/schema/zod/configPlatforms.ts
