+++ packages/core/src/schema/zod/configRootApp.ts
