+++ packages/core/src/schema/zod/configRootPlugin.ts
