+++ packages/core/src/schema/zod/configRootProject.ts
