+++ packages/core/src/schema/zod/configRootTemplates.ts
