+++ packages/core/src/schema/zod/electron/configPlatformElectron.ts
