+++ packages/core/src/schema/zod/ios/configPluginiOS.ts
