+++ packages/core/src/schema/zod/shared/configPluginShared.ts
