+++ packages/core/src/schema/zod/tizen/configPlatformTizen.ts
