+++ packages/core/src/schema/zod/windows/configPlatformWindows.ts
